netfilter:nf_tables MRF for nf_tables
Implemention of the MRF algortihm for the nf_tables firewall system.
CONFIG_SAL_GENERAL config flag needs to be enabled to activate the MRF
algorithm. To access important information about the alforithm
CONFIG_SAL_DEBUG is necessary. If CONFIG_SAL_MEM_HELPER_STRUCT flag
is activated the important information of a nft rule are extracted on
rule insertion and stored in a struct which holds the fields of
dst/src IP address, dst/src port and the protocol.

Signed-off-by: Jonas Köppeler <j.koeppeler@tu-berlin.de>